requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dnspython==2.4.1
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.101.0
fastapi-cache2==0.2.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
instagrapi==1.19.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.3
pendulum==2.1.2
Pillow==10.0.0
pycryptodomex==3.18.0
pydantic==1.10.9
pydantic-core==2.4.0
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
